Added tests for reserved task notification, new message related tests WIP
